backend/requirements.txt
Django==1.7.7
Markdown==2.6.1
django-filter==0.9.2
djangorestframework==3.1.1
mock==1.0.1
pep8==1.6.2
sanction==0.4.1
wsgiref==0.1.2
